requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

tqdm
sklearn
scikit-image
simpleitk
scipy
matplotlib

torchsummary
imageio
lmdb